WHENCE: nvidia: rearrange GSP-RM firmware lines
Group the lines for each version set of GSP-RM firmware files together,
so that they can be more easily removed or updated.

This commit makes no functional changes: the same files are placed in the
same locations, and the same links are created.

Nouveau/Nova use the GSP-RM firmware file, along with several auxilliary
firmware files, to support Nvidia Turing and later GPUs.  As new GPUs are
supported, newer versions of all files must be add to the repository.
Currently, two versions are present: 535.113.01 and 570.144.

To make it easier to add new versions, for distros to stop including older
versions, we regroup all of the relevant lines in the WHENCE file by
version, instead of by GPU.

Two things should be noted:

1. Some distros are apparently unable to handle changes to WHENCE where a
file is replaced with a symlink, or vice versa.

2. Most GPUs are just minor variants, and so these GPUs are represented
only by symlinks.

This means that care must be taken if a future firmware version includes
a new file for a GPU variant that was previously represented by a link.
Conversely, if a new version of file happens to be identical to another
file, it should not be replaced with a symlink.

Signed-off-by: Timur Tabi <ttabi@nvidia.com>