intel: qat: Fix missing link
When the QAT FW were in the root the qat_mmp.bin -> qat_895xcc_mmp.bin
was in place, that remains, but there isn't the equivilent of that in
the intel/qat directory so add in case the driver starts loading from
the new directory. Also group the links for this file close to the file.

Fixes: 06205c0e ("Move QAT firmware to intel/ subdirectory")
Signed-off-by: Peter Robinson <pbrobinson@gmail.com>